HFA: enable to read GCPs from PAM .aux.xml (fixes #4591)
The writing side is unchanged, that is GCPs are written to PAM, not as
internal GCPs.

In the process, use the more modern API for SRS, and move some code from
hfadataset.cpp to hfaopen.cpp, to avoid a hack in hfatest.cpp
