+++ gdal/frmts/hfa/hfatest.cpp
@@ -43,21 +43,6 @@ static void Usage()
 }
 
 /************************************************************************/
-/* Stub for HFAPCSStructToWKT, defined in hfadataset.cpp but used by    */
-/* hfaopen.cpp                                                          */
-/************************************************************************/
-#ifndef WIN32
-char *
-HFAPCSStructToWKT( const Eprj_Datum *,
-                   const Eprj_ProParameters *,
-                   const Eprj_MapInfo *,
-                   HFAEntry * )
-{
-    return nullptr;
-}
-#endif
-
-/************************************************************************/
 /*                                main()                                */
 /************************************************************************/
